vector_db_index/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.3.0
idna==3.4
importlib-metadata==6.7.0
newsapi-python==0.2.7
pymongo==4.6.0
python-dotenv==1.0.1
requests==2.31.0
sseclient-py==1.8.0
tabulate==0.9.0
together==0.2.6
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
